Feat tracing 0612 (#253)
* feat(tracing): 集成OpenTelemetry实现分布式追踪功能

- 添加core-tracing模块并集成OpenTelemetry自动配置
- 在各服务模块中引入OpenTelemetry依赖和配置
- 配置Jaeger和Loki容器用于链路追踪和日志收集
- 修改OrderDTO中seckillId字段类型从BigInteger改为Long
- 更新MongoDB配置和相关数据访问层代码适配Long类型
- 在各个应用的日志配置中添加OpenTelemetry appender支持
- 更新全局配置文件添加OTLP导出端点配置

* fix(loki): 配置Loki日志收集服务并修正OTLP端点

- 添加Loki配置文件挂载和数据持久化卷
- 配置Loki容器启动参数和网络设置
- 修正了日志收集系统的HTTP端点地址

* refactor(infrastructure): 重构监控栈配置

- 移除独立的 grafana 和 loki 服务
- 引入 grafana/otel-lgtm 镜像替代原有监控组件

* feat(gateway): 添加网关请求日志过滤器并优化跟踪配置

- 新增 RequestLoggingFilter 实现网关请求路径、状态码和耗时记录
- 配置 GatewayBootApplication 启用 REACTIVE Web应用类型
- 添加 Loki 日志聚合系统配置文件 loki.yml
- 更新 docker-compose.yml 中的 Grafana OTel LGTM 镜像配置
- 优化 OpenTelemetry 自动配置类注释和依赖注入
- 调整 SeckillService 实现类中的方法参数类型引用
- 将状态机服务中的日志级别从 info 调整为 debug
- 修复 goodskill-common.yml 配置文件中的中文注释乱码问题

* config: 更新应用配置和优化图片加载逻辑

- 在auth-service和ai服务中添加dev环境配置
- 新增logback-spring.xml日志配置文件用于统一日志管理
- 优化产品管理页面图片加载逻辑，支持HTTP URL直接显示
- 更新通用连接配置中的MongoDB和MinIO地址配置
- 修复公共服务扫描包配置问题
- 在秒杀响应监听器中添加延迟处理防止并发问题

* docs(readme): 更新文档添加OpenTelemetry集成说明

- 移除docker-compose.yml中的jaeger配置
- 添加OpenTelemetry可观测性能力说明
- 在依赖技术表格中增加OpenTelemetry条目
- 更新服务端口映射信息表
- 添加minio到hosts配置示例

* refactor(test): 优化测试代码中的数据类型和Mock配置

- 将BigInteger类型替换为long基本类型以简化测试数据设置
- 移除不必要的BigInteger导入语句
- 使用lenient()配置Mock行为以避免严格的验证异常
- 统一订单服务测试中的数据类型转换处理

---------

+++ goodskill-order-provider/goodskill-order-service/src/main/java/com/goodskill/order/entity/Order.java
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Data
@@ -23,7 +22,7 @@ public class Order {
     /**
      * 秒杀活动id
      */
-    private BigInteger seckillId;
+    private Long seckillId;
 
     /**
      * 用户手机号
